requirements.txt
# B站账号生命状态监控系统依赖包
# Python版本要求: >= 3.7

# 配置文件解析
PyYAML>=6.0

# HTTP请求（包含连接池和重试机制支持）
requests>=2.28.0
urllib3>=1.26.0

# 测试工具
pytest>=7.0
pytest-mock>=3.10

# 代码质量工具（开发用）
ruff>=0.1.0
mypy>=1.5

# 数据库（SQLite已内置，无需额外安装）

# 注意：
# - 如果使用QQ邮箱，需要获取授权码而不是QQ密码
# - 授权码获取方法：QQ邮箱 -> 设置 -> 账户 -> POP3/IMAP/SMTP服务 -> 生成授权码

